requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.9.1
dj-database-url==0.5.0
gunicorn==20.1.0
whitenoise==5.2.0
drf-yasg==1.20.0
coreapi==2.3.3